fix: url params setting & error page analytics
