+++ frontend/src/pages/Sessions/SessionsFeedV2/components/SessionsQueryBuilder/SessionsQueryBuilder.tsx
@@ -3,6 +3,7 @@ import {
 	useGetFieldTypesQuery,
 } from '@graph/hooks'
 import { SearchParamsInput } from '@graph/schemas'
+import { EmptySessionsSearchParams } from '@pages/Sessions/EmptySessionsSearchParams'
 import { useSearchContext } from '@pages/Sessions/SearchContext/SearchContext'
 import QueryBuilder, {
 	BOOLEAN_OPERATORS,
@@ -20,14 +21,45 @@ import QueryBuilder, {
 	VIEWED_BY_OPERATORS,
 } from '@pages/Sessions/SessionsFeedV2/components/QueryBuilder/QueryBuilder'
 import { useParams } from '@util/react-router/useParams'
+import { FieldArrayParam, QueryBuilderStateParam } from '@util/url/params'
+import { isEqual } from 'lodash'
 import moment from 'moment'
-import React from 'react'
+import React, { useEffect } from 'react'
+import {
+	ArrayParam,
+	BooleanParam,
+	JsonParam,
+	NumberParam,
+	StringParam,
+	useQueryParam,
+	useQueryParams,
+} from 'use-query-params'
 
 export const TIME_RANGE_FIELD: SelectOption = {
 	kind: 'single',
 	label: 'created_at',
 	value: 'custom_created_at',
 }
+
+const InitialSearchParamsForUrl = {
+	browser: undefined,
+	date_range: undefined,
+	device_id: undefined,
+	excluded_properties: undefined,
+	excluded_track_properties: undefined,
+	first_time: undefined,
+	hide_viewed: undefined,
+	identified: undefined,
+	length_range: undefined,
+	os: undefined,
+	referrer: undefined,
+	track_properties: undefined,
+	user_properties: undefined,
+	visited_url: undefined,
+	show_live_sessions: undefined,
+	environments: undefined,
+	app_versions: undefined,
+} as const
 
 const CUSTOM_FIELDS: CustomField[] = [
 	{
@@ -231,9 +263,138 @@ const SessionsQueryBuilder = React.memo(
 			variables: { project_id },
 		})
 
+		const searchContext = useSearchContext()
+
+		const {
+			searchParams,
+			setSearchParams,
+			segmentName,
+			page,
+			setPage,
+			selectedSegment,
+			setSelectedSegment,
+		} = searchContext
+
+		const [searchParamsToUrlParams, setSearchParamsToUrlParams] =
+			useQueryParams({
+				user_properties: FieldArrayParam,
+				identified: BooleanParam,
+				browser: StringParam,
+				date_range: JsonParam,
+				excluded_properties: FieldArrayParam,
+				hide_viewed: BooleanParam,
+				length_range: JsonParam,
+				os: StringParam,
+				referrer: StringParam,
+				track_properties: FieldArrayParam,
+				excluded_track_properties: FieldArrayParam,
+				visited_url: StringParam,
+				first_time: BooleanParam,
+				device_id: StringParam,
+				show_live_sessions: BooleanParam,
+				environments: ArrayParam,
+				app_versions: ArrayParam,
+				query: QueryBuilderStateParam,
+			})
+
+		const [activeSegmentUrlParam, setActiveSegmentUrlParam] = useQueryParam(
+			'segment',
+			JsonParam,
+		)
+
+		const [paginationToUrlParams, setPaginationToUrlParams] =
+			useQueryParams({
+				page: NumberParam,
+			})
+
+		useEffect(() => {
+			if (page !== undefined) {
+				setPaginationToUrlParams(
+					{
+						page: page,
+					},
+					'replaceIn',
+				)
+			}
+		}, [setPaginationToUrlParams, page])
+
+		useEffect(() => {
+			const areAnySearchParamsSet = !isEqual(
+				EmptySessionsSearchParams,
+				searchParams,
+			)
+
+			// Handles the case where the user is loading the page from a link shared from another user that has search params in the URL.
+			if (!segmentName && areAnySearchParamsSet) {
+				// `undefined` values will not be persisted to the URL.
+				// Because of that, we only want to change the values from `undefined`
+				// to the actual value when the value is different to the empty state.
+				const searchParamsToReflectInUrl = {
+					...InitialSearchParamsForUrl,
+				}
+				Object.keys(searchParams).forEach((key) => {
+					// @ts-expect-error
+					const currentSearchParam = searchParams[key]
+					// @ts-expect-error
+					const emptySearchParam = EmptySessionsSearchParams[key]
+					if (Array.isArray(currentSearchParam)) {
+						if (
+							currentSearchParam.length !==
+							emptySearchParam.length
+						) {
+							// @ts-expect-error
+							searchParamsToReflectInUrl[key] = currentSearchParam
+						}
+					} else if (currentSearchParam !== emptySearchParam) {
+						// @ts-expect-error
+						searchParamsToReflectInUrl[key] = currentSearchParam
+					}
+				})
+
+				setSearchParamsToUrlParams(
+					searchParamsToReflectInUrl,
+					'replaceIn',
+				)
+			}
+		}, [setSearchParamsToUrlParams, searchParams, segmentName])
+
+		useEffect(() => {
+			if (!isEqual(InitialSearchParamsForUrl, searchParamsToUrlParams)) {
+				setSearchParams(searchParamsToUrlParams as SearchParamsInput)
+			}
+			if (
+				paginationToUrlParams.page &&
+				page != paginationToUrlParams.page
+			) {
+				setPage(paginationToUrlParams.page)
+			}
+			// We only want to run this on mount (i.e. when the page first loads).
+			// eslint-disable-next-line react-hooks/exhaustive-deps
+		}, [])
+
+		// Session Segment Deep Linking
+		useEffect(() => {
+			if (selectedSegment && selectedSegment.id && selectedSegment.name) {
+				if (!isEqual(activeSegmentUrlParam, selectedSegment)) {
+					setActiveSegmentUrlParam(selectedSegment, 'replace')
+				}
+			} else if (activeSegmentUrlParam !== undefined) {
+				setActiveSegmentUrlParam(undefined, 'replace')
+			}
+			// eslint-disable-next-line react-hooks/exhaustive-deps
+		}, [selectedSegment, setActiveSegmentUrlParam])
+
+		useEffect(() => {
+			if (activeSegmentUrlParam) {
+				setSelectedSegment(activeSegmentUrlParam)
+			}
+			// We only want to run this on mount (i.e. when the page first loads).
+			// eslint-disable-next-line react-hooks/exhaustive-deps
+		}, [])
+
 		return (
 			<QueryBuilder
-				searchContext={useSearchContext()}
+				searchContext={searchContext}
 				timeRangeField={TIME_RANGE_FIELD}
 				customFields={CUSTOM_FIELDS}
 				fetchFields={fetchFields}
